+++ src/components/Navigation.vue
@@ -16,9 +16,9 @@
             <router-link class="nav-item" to="/npc" :class="$route.path === '/npc' ? 'current-page' : ''">
                 <span class="nav-link">NPC Generator</span>
             </router-link>
-<!--            <router-link class="nav-item" to="/crit">-->
-<!--                <span class="nav-link home">Critical Hit Generator</span>-->
-<!--            </router-link>-->
+            <router-link class="nav-item" to="/crit" :class="$route.path === '/crit' ? 'current-page' : ''">
+                <span class="nav-link home">Critical Hit Generator</span>
+            </router-link>
         </div>
     </div>
 </template>