+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.5.0
+- Critical Hit Generator added
+- Arrow recovery added
+
 ## 0.4.0
 - Fix Herbalism table column order and width
 - Npc Generator added
@@ -11,7 +15,7 @@
 - Add style to route if on route
 
 ## 0.2.1
-- Fix order of foudn herbs
+- Fix order of found herbs
 
 ## 0.2.0
 - Herbalism Generator added